fix BSResource and CRC
