+++ CommonLibF4/include/RE/Fallout.h
@@ -41,6 +41,10 @@
 #include "RE/Bethesda/BSPointerHandle.h"
 #include "RE/Bethesda/BSPreCulledObjects.h"
 #include "RE/Bethesda/BSResource.h"
+#include "RE/Bethesda/BSResource/BSResourceEnums.h"
+#include "RE/Bethesda/BSResource/LooseFileStreamBase.h"
+#include "RE/Bethesda/BSResource/Stream.h"
+#include "RE/Bethesda/BSResource/StreamBase.h"
 #include "RE/Bethesda/BSResourceNiBinaryStream.h"
 #include "RE/Bethesda/BSScaleformManager.h"
 #include "RE/Bethesda/BSScript.h"
@@ -129,7 +133,6 @@
 #include "RE/Bethesda/CombatFormulas.h"
 #include "RE/Bethesda/Console.h"
 #include "RE/Bethesda/ControlMap.h"
-#include "RE/Bethesda/ErrorCodes.h"
 #include "RE/Bethesda/Events.h"
 #include "RE/Bethesda/FavoritesManager.h"
 #include "RE/Bethesda/FormComponents.h"
@@ -168,8 +171,6 @@
 #include "RE/Bethesda/SendHUDMessage.h"
 #include "RE/Bethesda/Settings.h"
 #include "RE/Bethesda/SplineUtils.h"
-#include "RE/Bethesda/Stream.h"
-#include "RE/Bethesda/StreamBase.h"
 #include "RE/Bethesda/TESBoundAnimObjects.h"
 #include "RE/Bethesda/TESBoundObjects.h"
 #include "RE/Bethesda/TESCamera.h"
