+++ CommonLibF4/src/RE/Bethesda/BSResource.cpp
@@ -4,7 +4,7 @@ namespace RE
 {
 	namespace BSResource
 	{
-		BSResource::ErrorCode GetOrCreateStream(const char* a_fileName, BSTSmartPointer<BSResource::Stream>& a_result, bool a_writable, BSResource::Location* a_optionalStart)
+		ErrorCode GetOrCreateStream(const char* a_fileName, BSTSmartPointer<Stream>& a_result, bool a_writable, Location* a_optionalStart)
 		{
 			using func_t = decltype(&GetOrCreateStream);
 			REL::Relocation<func_t> func{ REL::ID(1143155) };
