+++ CommonLibF4/src/RE/Bethesda/CRC.cpp
